+++ src/components/TodoForm.jsx
@@ -1,12 +1,33 @@
-import React from 'react';
+import { React, useState } from 'react';
 import PropTypes from 'prop-types';
+import { v4 as uuidv4 } from 'uuid';
 import { InputGroup, FormControl, Button } from 'react-bootstrap';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { faPlus } from '@fortawesome/free-solid-svg-icons';
+import db from '../utils/db';
 
-const TodoForm = React.memo((props) => {
+const TodoForm = (props) => {
 
-  const { todo, inputTodoHandler, addTodoListHandler } = props;
+  const { todoList, setTodoList } = props;
+
+  const [todo, setTodo] = useState('');
+
+  const inputTodoHandler = (event) => {
+    setTodo(event.target.value);
+  };
+
+  const addTodoListHandler = async() => {
+    const newTodoList = [...todoList];
+    const newTodo = {
+      id: uuidv4(),
+      text: todo,
+      done: false,
+    };
+    newTodoList.push(newTodo);
+    setTodoList(newTodoList);
+    setTodo('');
+    await db.todoList.add(newTodo);
+  };
 
   return (
     <InputGroup>
@@ -23,12 +44,11 @@ const TodoForm = React.memo((props) => {
       </InputGroup.Append>
     </InputGroup>
   );
-});
+};
 
 TodoForm.propTypes = {
-  todo: PropTypes.string,
-  inputTodoHandler: PropTypes.func,
-  addTodoListHandler: PropTypes.func,
+  todoList: PropTypes.array,
+  setTodoList: PropTypes.func,
 };
 
 export default TodoForm;
